refs #appdev Typescript convertion commit
