+++ .gitignore
@@ -28,6 +28,8 @@ coverage
 # Grunt intermediate storage (https://gruntjs.com/creating-plugins#storing-task-files)
 .grunt
 
+lib/*
+
 # Bower dependency directory (https://bower.io/)
 bower_components
