--- app/models/login_model.php
@@ -1,10 +1,0 @@
-const mongoose = require('mongoose');
-
-const NoteSchema = mongoose.Schema({
-    title: String,
-    content: String
-}, {
-    timestamps: true
-});
-
-module.exports = mongoose.model('Note', NoteSchema);
