--- config/database.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-    url: 'mongodb://localhost:27017/easy-notes'
-}
