+++ package.json
@@ -2,10 +2,10 @@
   "name": "app-service",
   "version": "1.0.0",
   "description": "",
-  "main": "server.js",
+  "main": "lib/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start":"node server.js"
+    "build": "tsc",
+    "start": "node lib/index.js"
   },
   "keywords": [
     "Express",
@@ -14,11 +14,12 @@
     "Mongoose",
     "Notes"
   ],
-  "author": "aravind",
-  "license": "MIT",
+  "author": "karthikeyan",
   "dependencies": {
     "body-parser": "^1.19.0",
     "express": "^4.17.1",
-    "mongoose": "^5.8.3"
+    "mongodb": "^3.4.1",
+    "mongoose": "^5.8.3",
+    "typescript": "^3.7.4"
   }
 }
