+++ src/app/routes/Router.ts
@@ -1,0 +1,11 @@
+import LoginController from '../controllers/LoginController';
+
+export default class Router {
+
+    public static listenMethods = (app)=>{
+        app.post('/signup', LoginController.createUsers);
+        app.post('/signin', LoginController.loginAuth);
+
+    }
+}
+
