+++ src/config/dbConfig.ts
@@ -1,0 +1,3 @@
+export default {
+    url: 'mongodb://localhost:27017/app-service'
+}
